feat: scale menu font size across intro, exit, and options layers
Replaces hardcoded 48px font sizes in ExitLayer and OptionLayer with
menuFontSizeForWidth, matching the existing IntroLayer behaviour.
Updates breakpoints to 24px (<1280), 32px (<1920), and 48px (>=1920)
for a noticeable jump between common resolutions. Swaps 40px out of
the atlas for 24px to match the new size set.

+++ game/src/layer/introlayer.cpp
@@ -89,11 +89,10 @@ void IntroLayer::onAttach() {
 
     auto makeMenuButton = [fontNameStr](std::string_view message) {
         return std::make_unique<ui::Button>(ui::ButtonCreateInfo{
-            .topLeft     = glm::vec2{ 0.F },
-            .bottomRight = glm::vec2{ 0.F },
-            .message     = std::string(message),
-            .fontSize    = ui::menuFontSizeForWidth(
-                static_cast<uint32_t>(orthoCamera->getWidth())),
+            .topLeft      = glm::vec2{ 0.F },
+            .bottomRight  = glm::vec2{ 0.F },
+            .message      = std::string(message),
+            .fontSize     = ui::menuFontSizeForWidth(orthoCamera->getWidth()),
             .fontName     = fontNameStr,
             .buttonColor  = buttonColor,
             .textColor    = textColor,
